requirements.txt
python-telegram-bot==21.0.1
yandexcloud
boto3==1.34.0
python-dotenv==1.0.0
pydub==0.25.1
requests==2.31.0